consolidate all cobra usage into HandleFGCobraSearch; define line between querysupport and connection packages
